feat: anchor the detail ladder physically and open the player-view zoom
User calibration during Task 3 acceptance: the 3 px pixel target pinned
leaf primitives below visibility at every zoom, so the mosaic read as a
continuous field. The selection now works on a physical ladder — levels
halve the spec 5 primitive edge from the cell spacing down to 10 m
(draft 14 levels, standard and high 13), each level engaging when about
96 primitives span the viewport (one level per zoom octave) — and the
camera keeps zooming past the ladder floor to a ~100 m player view
(map max zoom 4096 -> 409,600, globe max scale 64 -> 131,072), where
the 10 m data atoms render as plainly visible solid triangles.

Deep zooms sit far beyond f32 uniform precision, so detail geometry is
camera-relative: amplified meshes and river polylines stay f64 to the
upload, rebase there about stored anchors (map projection-plane view
center, globe front direction), and a second detail_transform uniform
folds the anchor into the translation column in f64 — identical scale
columns keep horizon and silhouette math exact, and zero anchors reduce
it to the main transform. Rivers get their own vertex entries sharing
the factored overlay body. Globe trackball drags divide their angle by
the orthographic scale past 1 so deep close-ups pan under the cursor
instead of flinging. The authoritative cell-view pipelines stay on
absolute coordinates: beyond the old 4096x they show gradual vertex
crawl, recorded as known debt for the data-inspection path.

Verified live on the standard tier: shoreline gradients show the bare
atom mosaic at depth, rivers stay pinned through pans, rendering is
crawl-free at maximum zoom, and the deep globe stays steerable. Gates:
fmt, clippy -D warnings, wasm check, full release regression 180
targets / 1,532 passed / 0 failed.

+++ src/view/spherical_mesh.rs
@@ -1314,11 +1314,12 @@ fn append_edge_segment(
 
 /// One direction-domain amplified subdivision mesh shared by both presenters.
 ///
-/// Directions are unit vectors and colors are pre-lit sRGB bytes; the mesh is
-/// built once per published world and re-projected per map view.
+/// Directions are f64 unit vectors (full precision until the GPU rebase
+/// subtracts the camera anchor) and colors are pre-lit sRGB bytes; the
+/// mesh is rebuilt per detail selection and re-projected per map view.
 #[derive(Debug, Clone)]
 pub struct AmplifiedSurfaceMesh {
-    directions: Vec<[f32; 3]>,
+    directions: Vec<[f64; 3]>,
     colors: Vec<[u8; 4]>,
     indices: Vec<u32>,
 }
@@ -1326,7 +1327,7 @@ pub struct AmplifiedSurfaceMesh {
 impl AmplifiedSurfaceMesh {
     /// Validates cardinalities, finiteness, and index bounds once.
     pub fn new(
-        directions: Vec<[f32; 3]>,
+        directions: Vec<[f64; 3]>,
         colors: Vec<[u8; 4]>,
         indices: Vec<u32>,
     ) -> Result<Self, SphericalMeshError> {
@@ -1355,7 +1356,7 @@ impl AmplifiedSurfaceMesh {
     }
 
     /// Returns the unit directions of all subdivision vertices.
-    pub fn directions(&self) -> &[[f32; 3]] {
+    pub fn directions(&self) -> &[[f64; 3]] {
         &self.directions
     }
 
@@ -1379,9 +1380,9 @@ impl AmplifiedSurfaceMesh {
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub struct RiverPolylineSegment {
     /// The upstream cell centroid direction.
-    pub start: [f32; 3],
+    pub start: [f64; 3],
     /// The downstream cell centroid direction.
-    pub end: [f32; 3],
+    pub end: [f64; 3],
     /// The published Strahler order driving the symbolic line width.
     pub strahler_order: u8,
 }
@@ -1392,23 +1393,26 @@ pub struct RiverPolylineSegment {
 /// concern (amplified triangles re-cut, river reaches simply drop).
 pub fn project_unit_direction(
     projection: SphericalProjection,
-    direction: [f32; 3],
-) -> Option<[f32; 2]> {
+    direction: [f64; 3],
+) -> Option<[f64; 2]> {
     let [x, y, z] = direction;
-    let longitude = f64::from(y).atan2(f64::from(x));
-    let latitude = f64::from(z).clamp(-1.0, 1.0).asin();
+    let longitude = y.atan2(x);
+    let latitude = z.clamp(-1.0, 1.0).asin();
     let relative = wrap_radians(longitude - projection.central_meridian());
     projection
         .forward_latitude_relative_longitude(latitude, relative)
         .ok()
-        .map(|point| [point.x() as f32, point.y() as f32])
+        .map(|point| [point.x(), point.y()])
 }
 
 /// One projected amplified map vertex ready for GPU packing.
+///
+/// The position stays f64 so the renderer's camera-relative rebase
+/// subtracts the anchor before any f32 quantization.
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub struct AmplifiedMapVertex {
     /// The projected planar position.
-    pub position: [f32; 2],
+    pub position: [f64; 2],
     /// The pre-lit sRGB color carried over from the direction-domain vertex.
     pub color: [u8; 4],
 }
@@ -1429,7 +1433,7 @@ pub fn project_amplified_map(
     use rayon::prelude::*;
 
     let bounds = projection.bounds();
-    let half_width = ((bounds.max_x() - bounds.min_x()) * 0.5 + SPAN_EPSILON) as f32;
+    let half_width = (bounds.max_x() - bounds.min_x()) * 0.5 + SPAN_EPSILON;
     let central_meridian = projection.central_meridian();
     #[cfg(not(target_arch = "wasm32"))]
     let vertex_iter = mesh.directions().par_iter().zip(mesh.colors().par_iter());
@@ -1437,7 +1441,7 @@ pub fn project_amplified_map(
     let vertex_iter = mesh.directions().iter().zip(mesh.colors().iter());
     let mut vertices: Vec<AmplifiedMapVertex> = vertex_iter
         .map(|(direction, color)| AmplifiedMapVertex {
-            position: project_unit_direction(projection, *direction).unwrap_or([f32::NAN; 2]),
+            position: project_unit_direction(projection, *direction).unwrap_or([f64::NAN; 2]),
             color: *color,
         })
         .collect();
@@ -1451,8 +1455,8 @@ pub fn project_amplified_map(
         let span = positions
             .iter()
             .map(|p| p[0])
-            .fold(f32::NEG_INFINITY, f32::max)
-            - positions.iter().map(|p| p[0]).fold(f32::INFINITY, f32::min);
+            .fold(f64::NEG_INFINITY, f64::max)
+            - positions.iter().map(|p| p[0]).fold(f64::INFINITY, f64::min);
         if finite && span <= half_width {
             indices.extend_from_slice(triangle);
         } else {
@@ -1481,7 +1485,7 @@ fn append_seam_cut_triangle(
     let mut recovered = Vec::with_capacity(TRIANGLE_VERTEX_COUNT);
     for &corner in &corners {
         let [x, y, z] = mesh.directions()[corner as usize];
-        let Ok(direction) = UnitVector3::new(f64::from(x), f64::from(y), f64::from(z)) else {
+        let Ok(direction) = UnitVector3::new(x, y, z) else {
             return;
         };
         recovered.push(direction);
@@ -1516,7 +1520,7 @@ fn append_seam_cut_triangle(
         };
         for (point, _direction) in &projected {
             vertices.push(AmplifiedMapVertex {
-                position: [point.x() as f32, point.y() as f32],
+                position: [point.x(), point.y()],
                 color: flat_color,
             });
         }
